+++ native/zowed/schemas/responses.hpp
@@ -254,6 +254,11 @@ ZJSON_SCHEMA(ChtagFileResponse,
     FIELD_REQUIRED(success, BOOL)
 );
 
+struct CopyUssResponse {};
+ZJSON_SCHEMA(CopyUssResponse,
+    FIELD_REQUIRED(success, BOOL)
+);
+
 struct CreateFileResponse {};
 ZJSON_SCHEMA(CreateFileResponse,
     FIELD_REQUIRED(success, BOOL)